# feat: Faster string serialization by escaping in byte space.