During linux `perf` profiling of the python client, I've noticed that the `write_escaped_unquoted` function was on the hot path.

![profile_graph (1)](https://user-images.githubusercontent.com/1499096/204813453-563552c9-0b1e-4f76-b47d-737e6f36ca03.svg)

This PR improves its performance by around ~31%. 

It does so by using unsafe APIs for manipulating strings as vectors: Character escaping is done as bytes rather than chars.

We also get some extra speed by avoiding a branch in the escaping loop by unsafely setting the vector length and setting its contents rather than calling `.push` repeatedly.

We got an extra bit of performance by inlining two functions that get called repeatedly.

Profiling code:

```python
    def test_string_escaping_10m(self):
        count = 10_000_000
        slist = [f's={i:09}==abc \\' for i in range(count)]
        series = pd.Series(slist, dtype='string[pyarrow]')
        df = pd.DataFrame({
            'col1': series,
            'col2': series,
            'col3': series,
            'col4': series,
            'col5': series,
            'col6': series})
        
        buf = qi.Buffer()

        # Warm up and pre-size buffer
        buf.pandas(df, table_name='tbl1', symbols=True)
        buf.clear()

        # Run
        t0 = time.monotonic()
        buf.pandas(df, table_name='tbl1', symbols=True)
        t1 = time.monotonic()
        print(f'Time: {t1 - t0}, size: {len(buf)}')
```

Perf difference:
* With the old code, three sample runs took: 4.12s, 4.01s, 4.07s (avg: 4.06s).
* With the code in this PR, three sample runs took: 2.74s, 2.79s, 2.83s (avg: 2.78s).